feat(FX-4478): Re-position elements in the menu (mWeb)
